requirements.txt
torch
torchvision
streamlit
opencv-python
facenet-pytorch
Pillow
tqdm
pandas
time